Extract out tab component
