+++ src/components/admin/TabBar/index.tsx
@@ -1,0 +1,1 @@
+export { TabBar } from "./TabBar";
