+++ src/components/admin/WorldNav/WorldNav.tsx
@@ -1,6 +1,5 @@
 import React, { useMemo } from "react";
-import { Link } from "react-router-dom";
-import classNames from "classnames";
+import { TabBar } from "components/admin/TabBar";
 
 import { ADMIN_IA_WORLD_EDIT_PARAM_URL } from "settings";
 
@@ -8,36 +7,28 @@ import { generateUrl } from "utils/url";
 
 import { useWorldParams } from "hooks/worlds/useWorldParams";
 
-import * as TW from "./WorldNav.tailwind";
 import { WorldNavLabelMap } from "./WorldNavLabelMap";
 
 export const WorldNav: React.FC = () => {
   const { worldSlug, selectedTab } = useWorldParams();
 
-  const renderedTabs = useMemo(() => {
-    return Object.entries(WorldNavLabelMap).map(([key, label]) => {
-      const url = generateUrl({
-        route: ADMIN_IA_WORLD_EDIT_PARAM_URL,
-        required: ["worldSlug", "selectedTab"],
-        params: { worldSlug, selectedTab: key },
-      });
+  const tabs = useMemo(() => {
+    return Object.fromEntries(
+      Object.entries(WorldNavLabelMap).map(([key, label]) => {
+        const url = generateUrl({
+          route: ADMIN_IA_WORLD_EDIT_PARAM_URL,
+          required: ["worldSlug", "selectedTab"],
+          params: { worldSlug, selectedTab: key },
+        });
 
-      const classes = classNames(TW.tab, {
-        [TW.selectedTab]: selectedTab === key,
-        [TW.notSelectedTab]: selectedTab !== key,
-      });
-
-      return (
-        <Link key={key} to={url} className={classes}>
-          {label}
-        </Link>
-      );
-    });
-  }, [selectedTab, worldSlug]);
+        return [key, { url, label }];
+      })
+    );
+  }, [worldSlug]);
 
   return (
     <div className="AdminVenueView__options -mb-px flex bg-white shadow">
-      {renderedTabs}
+      <TabBar tabs={tabs} selectedTab={selectedTab} />
     </div>
   );
 };
